requirements.txt

flask
flask-cors
python-dotenv
google-generativeai
google-adk
requests
beautifulsoup4
aiohttp
colorama
